Plot feature added
Now PlotVariableOneEnergy can plot any variable from a result root file, by introducing the name of the variable, the conditions, nbins, etc. As if you plot in ROOT Browser a variable. But in this case, the function is able to plot one energy.  
Now only works for one energy, I am improving the code for all the energies on demand.